undirected run config generator (#427)
* checkpoint

* refactor model config generation

* more cleanup

* More improvements

* refactor main step function

* refactor coordinate picking code

* add assert and rename variable

* refactor determine if done

* Update comments. Only create neighborhood as needed

* Extract neighborhoodConfig from SearchConfig

* minor tweaks

+++ model_analyzer/config/generate/search_dimension.py
@@ -65,6 +65,6 @@ class SearchDimension:
         if self._type == SearchDimension.DIMENSION_TYPE_LINEAR:
             return idx + 1
         elif self._type == SearchDimension.DIMENSION_TYPE_EXPONENTIAL:
-            return math.pow(2, idx)
+            return int(math.pow(2, idx))
         else:
             raise Exception(f"Unknown type {self._type}")
